add text_jp
